rename authtools.py => auth.py
